[MCUX-83803][MISRA] SDK components: fix MISRA C-2012 violations
Changes per file:
- internal_flash/fsl_adapter_k4_flash.c:
  - Add explicit 'void' parameter to HAL_FlashInit() (MISRA Rule 17.3)
  - Cast memcpy() return values to void (MISRA Rule 17.7)
  - Remove dead assignment '*sav_cfg = 0U' before conditional init
  - Replace ternary bool expression with explicit bitmask comparison
    in HAL_FlashEccStatusRaised() (MISRA Rule 14.4)

- mem_manager/fsl_component_mem_manager_light.c:
  - Add explicit type casts (uint8_t, uint32_t, size_t) in integer
    comparisons to fix essential type mismatch (MISRA Rule 10.4)

- osa/fsl_os_abstraction_threadx.c:
  - Replace integer literals 0/1 with unsigned suffixed forms 0U/1U
    (MISRA Rule 7.2)
  - Add (uint32_t) casts to sizeof() expressions in arithmetic
    (MISRA Rules 10.4, 18.4)

- osa/fsl_os_abstraction_threadx.h:
  - Add explicit (UINT) and (osa_task_priority_t) casts in
    PRIORITY_OSA_TO_RTOS and PRIORITY_RTOS_TO_OSA macros
    (MISRA Rule 10.1/10.4)

- serial_manager/fsl_component_serial_manager.c:
  - Change '== 0' to '== 0U' for unsigned integer comparison
    (MISRA Rule 7.2)

- uart/fsl_adapter_lpuart.c:
  - Cast LPUART_Deinit() return value to void (MISRA Rule 17.7)

+++ mem_manager/fsl_component_mem_manager_light.c
@@ -561,7 +561,7 @@ mem_status_t MEM_RegisterExtendedArea(memAreaCfg_t *area_desc, uint8_t *p_area_i
                     st = kStatus_MemInitError; /* overlap with previously registered blocks */
                     break;
                 }
-                assert(id < UINT8_MAX);
+                assert(id < (uint8_t)UINT8_MAX);
                 id++;
             }
             if (st != kStatus_MemSuccess)
@@ -943,7 +943,7 @@ static void *MEM_BufferAllocate(uint32_t numBytes, uint8_t poolId)
                     break;
                 }
             }
-            assert(area_id < UINT8_MAX);
+            assert(area_id < (uint8_t)UINT8_MAX);
             area_id++;
         }
     }
@@ -965,7 +965,7 @@ void *MEM_BufferAllocWithId(uint32_t numBytes, uint8_t poolId)
 #endif
     void_ptr_t buffer_ptr;
 
-    if (numBytes > UINT16_MAX)
+    if (numBytes > (uint32_t)UINT16_MAX)
     {
         return NULL;
     }
@@ -1262,7 +1262,7 @@ void *MEM_BufferRealloc(void *buffer, uint32_t new_size)
     uint16_t block_size  = 0U;
     do
     {
-        if (new_size > UINT16_MAX)
+        if (new_size > (uint32_t)UINT16_MAX)
         {
             realloc_buffer = NULL;
             /* Bypass he whole procedure so keep original buffer that cannot be reallocated */
@@ -1378,7 +1378,7 @@ void *MEM_CallocAlt(size_t len, size_t val)
 {
     size_t blk_size;
 
-    if (len > UINT16_MAX / val)
+    if (len > (size_t)(UINT16_MAX / val))
     {
         return NULL;
     }
